Redo DnD to support dragging from library into place in the graph
It's kindof messy but it works for now.  This proves that it's OK
to change DragHandler.target during a drag, which is kindof cool.
But actual DnD via the Drag attached property interferes with this
technique: after the Drag starts, I haven't found a way to make it
transition to dragging an actual Item instead.  So it's easier to
avoid that and set the DH target after the drag starts (drag
threshold is exceeded).